Add mint-test-pr script for quickly testing pull requests from github
The script is called as `mint-test-pr <project> <pull request number>`
It clones the given project, installs build deps, checks out the
branch of the given pr, builds it, and then installs the generated
debs. If a folder is supplied as a fourth arg, the debs will be
copied to that folder instead of being installed. Everything is done
in a temp folder and then removed upon completion so as to avoid
potential conflicts with other files and folders on the system.