Override dpkg-deb to force XZ compression
Packages which architecture is "all", or which do not require
compilation/linking against a particular target can be fully compatible
between Mint and LMDE and do not require specific builds.

In Ubuntu 22.04 however, and thus Mint 21, dpkg-deb switched from XZ
to ZSTD compression. This is a problem because Debian is yet to support
ZSTD decompression in dpkg, so packages built on a Mint 21 host cannot
be decompressed in LMDE.

Until Debian brings ZSTD support to dpkg we need to stick to XZ when
building cross-distro packages.

This /usr/local/bin/dpkg-deb wrapper overrides the default compression
and sets it to XZ when we use dpkg-buildpackage.